Atualizacao do Novo Design     -Atualização nova tela e nova estrutura     -Ajuste Readme     -Refatoração e reestruturação das classes dos elementos     -Inserção do módulo para inserir comentarios ao inserir glicemias
