+++ ContagemInsulina/Config.cs
@@ -83,8 +83,8 @@ namespace ContagemInsulina
 
         private void Config_FormClosing(object sender, FormClosingEventArgs e)
         {
-            Form1 f1 = new Form1();
-            f1.Show();
+           // Form1 f1 = new Form1();
+            //f1.Show();
         }
 
         private void button1_Click(object sender, EventArgs e)
