+++ ContagemInsulina/Program.cs
@@ -16,7 +16,7 @@ namespace ContagemInsulina
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new FormMenu());
         }
     }
 }
