+++ README.md
@@ -1,9 +1,42 @@
-# ContagemInsulina
+# Contagem de Insulina
 
-![V1](assets/Tela1.png)
-/n
+#### _Para um melhor controle e gestao da insulina a ser aplicada_
+
+![](https://img.shields.io/github/stars/pandao/editor.md.svg) ![](https://img.shields.io/github/tag/pandao/editor.md.svg) ![](https://img.shields.io/github/release/pandao/editor.md.svg) ![](https://img.shields.io/github/issues/pandao/editor.md.svg)
+
+Contagem de insulina trata de um software para auxiliar na correcao, contagem de carboidratos e analise para realizacao de exercicio fisico, pensando na melhor qualidade de vida para o utilizado.
+
+- Auxilia de forma rapida e pratica a saber a quantidade de insulina a ser aplicada.
+- Permite realizar uma analise se é viavel ou não realizar exercicio fisico.
+- Possibilidade acompanhar o desenvolvimento glicemico do usuario e muito mais.
+
+## Funcionalidades
+
+- Analises possiveis para a glicemia: Correcao, Alimentacao(Contagem de carboidratos) e Exercicio fisico
+- Sistema monitorando o momento do dia para sugerir a alimententacao atual
+- Analise geral da acao a ser tomada a partir do valor da glicose e parametros escolhidos para analise
+- Possibilidade de inserir todos os parametros como: Fator sensibilidade, Glicemia alvo, Relacao carboidrato, etc
+- Possibilidade de ajustar a quantidade de carboidratos por refeicao
+
+### Autor
+
+- Edgar Oliveira Cardoso
+
+### Diagramas representando a funcionalidade do sistema:
+
 Fluxograma
-![V1](assets/Fluxograma.png)
-/n
+
+![](assets/Fluxograma.png)
+
 Diagrama de Caso de Uso
-![V1](assets/DiagramaCasoDeUso.png)
+
+![](assets/DiagramaCasoDeUso.png)
+
+### Versoes
+
+> Versao 1.0
+> Versao inicial, contando com as funcionalidades ativas e garantindo um controle eficaz da glicemia
+
+Imagens:
+
+![](assets/V1/Tela1.png)
